requirements.txt
anthropic>=0.40.0
py-clob-client==0.19.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
httpx==0.27.2
python-dotenv==1.0.1
scipy==1.14.1
numpy==2.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0
aiofiles==24.1.0
